% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_count_follower}
\alias{get_count_follower}
\title{Get follower count from a user page}
\usage{
get_count_follower(userpage = NULL)
}
\arguments{
\item{userpage}{A user page. An output from \code{\link{get_page_user}}.}
}
\value{
Returns an integer follower count or NA.
}
\description{
Parses a text userpage and extracts follower count
}
\details{
If the parsing fails, NA is returned with a warning.
}
\examples{
pu <- get_page_user("instagram")
cf <- get_count_follower(pu)
}
