% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_page_hashtag}
\alias{get_page_hashtag}
\title{Get hashtag page}
\usage{
get_page_hashtag(hashtag = NULL)
}
\arguments{
\item{hashtag}{A character denoting a valid instagram hashtag.}
}
\value{
Returns an httr webpage object as text or NA.
}
\description{
Gets the hashtag page
}
\details{
If the tag is not valid and/or if it does not return a webpage,
NA is returned with a warning.
}
\examples{
# get page for hashtag: instagram
un <- get_page_hashtag("instagram")
}
\seealso{
\code{\link{get_count_hashtag}}
}
