% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_page_info}
\alias{get_page_info}
\title{Get user page info}
\usage{
get_page_info(username = NULL)
}
\arguments{
\item{username}{A character vector of one or more valid instagram usernames.}
}
\value{
Returns a dataframe with post counts, follower count and following count.
}
\description{
Gets the three metrics from a user page
}
\details{
If the username is not valid, an empty data.frame is returned
}
\examples{
# get page for username: instagram
pu <- get_page_info("instagram")
}
\seealso{
\code{\link{get_count_post}}, \code{\link{get_count_follower}}, \code{\link{get_count_following}}
}
