% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_page_user}
\alias{get_page_user}
\title{Get user page}
\usage{
get_page_user(username = NULL)
}
\arguments{
\item{username}{A character denoting a valid instagram username.}
}
\value{
Returns an httr webpage object text or NA.
}
\description{
Gets the user page
}
\details{
If the username is not valid, it does not return a webpage,
NA is returned with a warning.
}
\examples{
# get page for username: instagram
pu <- get_page_user("instagram")
}
\seealso{
\code{\link{get_count_post}}, \code{\link{get_count_follower}}, \code{\link{get_count_following}}
}
