% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_count_hashtag}
\alias{get_count_hashtag}
\title{Get hashtag count from a hashtagpage}
\usage{
get_count_hashtag(hashtagpage = NULL)
}
\arguments{
\item{hashtagpage}{A hashtagpage. An output from \code{\link{get_page_hashtag}}.}
}
\value{
Returns a numeric follower count or NA.
}
\description{
Parses a text hashtagpage and extracts hashtag count
}
\details{
If the parsing fails, NA is returned with a warning.
}
\examples{
ph <- get_page_hashtag("instagram")
ch <- get_count_hashtag(ph)
}
